project completed

+++ app/Http/Controllers/ListingController.php
@@ -51,6 +51,8 @@ class ListingController extends Controller
             $formFields['logo']=$path;
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message','Listing created successfully');
@@ -63,6 +65,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing){
+
+        //Make sure logged in user is owner
+        if($listing->user_id !=auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         
         $formFields = $request->validate([
             
@@ -87,7 +94,16 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+         //Make sure logged in user is owner
+         if($listing->user_id!=auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', "Deleted successfully");
+    }
+
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
